Make pragmas context-sensitive, improve =include.
Pragmas can now be context-sensitive.  That is,
they can know where they're being executed
(via perky.breadcrumbs), and can modify the
root or the leaf or whatever they like.

pragma_include() now modifies the current leaf
rather than always merging into the root.  Uses
a new sophisticated merge_dict_and_list()
function in utilities.

Changed Perky behavior: previously you were permitted
to redefine a key in a dict context as often as you
like, e.g.:
   a = 3
   a = 5
This is no longer legal, because it seems like it
would lead to errors.  Note however that this is
explicitly per-text (or, if you like, per-file).
Redefining a key in the current context from an
=include'd file is explicitly *permitted*.

This checkin also removes the include() and includes()
functions, as they were always kind of a 'code smell'.