Pragma now accepts quoted strings.  0.2!
The pragma line parser now uses the Perky parser
to parse the argument to the pragma, so it now
supports quoted strings.  It also throws an
exception if you try to use any of the Perky
special values:

    =include {
    =include [
    =include '''
    etc.

(This leaves open the possible future feature
where you can specify an arbitrary Perky value
as a pragma argument, like a list or a dict.
I don't need it yet, and I'm not sure I want it,
so right now it's an error.)

I'm also going to release this as version 0.2.
Perky is coming along, slowly but surely!