0.5.5.  Bugfix, serializing special characters.
If you called perky.loads() on this text:

    Apple ][ = {
        bits = 8
    }

it fails, because Perky doesn't like those square
brackets there.  It wants them quoted.  I'm not
sure whether or not that's the behavior I want.
But at the very least, when Perky *serializes*
the equivalent dict:
    { "Apple ][" : { "bits" : 8 } }
it should jolly well emit text that it would
happily parse.

It could be that Perky *should* parse the above
text.  But that's a decision for another day.