0.6.1!  Fixed pushback_str_iterator.__bool__.
For 0.6, I fiddled with
pushback_str_iterator.__bool__,
and completely broke it.

First, __bool__ has to return a Bool,
it can't retrun an arbitrary object
representing truthyness or falseyness.

Second, the logic in my new __bool__ was
wrong, because it didn't handle a boundary
condition.  If the stack was empty, and
the iterator was exhausted *but we hadn't
noticed yet*, it would have returned True.
If the stack is empty and we still have an
iterator, we need to test that the iterator
really returns a value.

This didn't actually break Perky, because
it never tests the truthyness of pushback
iterators--it simply iterates over them, and
exits the loop through the normal StopIteration
mechanism.  All that code was unaffected.
And since pushback_str_iterator is an internal
data structure, not supported for external
callers, this bug should have affected
literally nobody.

(Maybe this is an argument for removing __bool__,
as nobody's calling it.  I'm leaving it for now,
as I've gone to the effort of fixing it,
adding a regression test, etc.)