0.6.2!  Updated copyright year, fixed 3.7- compat.
Updated the copyright year to 2023 for all files.

During development of 0.6, I added an equals-sign-in-fstring
to the test suite at the end of utility.py.  This broke
compatibility with versions of Python prior to 3.8.
So, removed the equals sign, and ran the test with all
versions of Python from 3.6 to 3.11 inclusive, and all
is well again.